fix(settlement): retry request timeout errors
Alloy's default policy does not retry HTTP 408 responses, allowing
provider request timeouts to reach the non-recoverable settlement path.

Retry standardized HTTP 408 responses. For direct JSON-RPC errors,
require both the implementation-defined -32009 code and tight timeout
wording so unrelated permanent errors remain non-transient. Keep
task-level restart behavior unchanged.

Move utility tests into a dedicated module and cover the captured
error shape plus false-positive cases.

The following verification passed:

- `cargo check --workspace --tests --all-features`
- `mdbook build docs/knowledge-base/`
- `cargo make ci-all`
- `cargo nextest run --workspace` (950 tests)
